+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### CHANGED
+
+- Minimum Python version from 3.9 to 3.10 (Required to build v0.4.0 of our docker image).
+
 ## [0.4.0] - 2025-11-10
 
 ### Changed